refactoring, half-baked drag bar
