+++ index.html
@@ -20,7 +20,9 @@
 <script type="text/javascript" language="javascript" src="js/audio.js"></script>
 <script type="text/javascript" language="javascript" src="js/lib/lz-string.js"></script>
 <script type="text/javascript" language="javascript" src="js/ui.js"></script>
-<script type="text/javascript" language="javascript" src="js/library.js"></script>
+<script type="text/javascript" language="javascript" src="js/ui-mixer.js"></script>
+<script type="text/javascript" language="javascript" src="js/ui-playlist.js"></script>
+<script type="text/javascript" language="javascript" src="js/ui-library.js"></script>
 <script type="text/javascript" language="javascript">
