+++ js/ui.js
@@ -348,6 +348,58 @@ class NoteAction extends Action {
 	}
 }
 
+class PlaybackMarker {
+    constructor() {
+        this.measure = null;
+        this.time = -1;
+        this.handleWidth = 10;
+        this.buildUI();
+    }
+
+    buildUI() {
+        // just a thin div
+        this.playbackBox = document.createElement("div");
+        // absolutely positioned
+        this.playbackBox.className = "playbackBox";
+        this.playbackBox.style.width = "4px";
+        this.playbackBox.style.height = (gridSizeY * (13 + 1)) + "px";
+        this.playbackBox.style.left = 0;
+        this.playbackBox.style.top = 0;
+
+        this.playbackHandle = document.createElement("div");
+        // absolutely positioned
+        this.playbackHandle.className = "playbackHandle";
+        this.playbackHandle.style.width = (4 + this.playbackHandleWidth*2) + "px";
+        this.playbackHandle.style.height = (gridSizeY * (13 + 1)) + "px";
+        this.playbackHandle.style.left = -this.playbackHandleWidth;
+        this.playbackHandle.style.top = 0;
+        // back-references
+        this.playbackBox.timingBar = this;
+        this.playbackHandle.timingBar = this;
+    }
+
+    setTime(measure, time) {
+        if (measure != this.measure) {
+            if (this.measure != null) {
+                this.playbackBox.remove();
+                this.playbackHandle.remove();
+            }
+            this.measure = measure;
+            this.measure.timingContainer.appendChild(this.playbackBox);
+            this.measure.timingContainer.appendChild(this.playbackHandle);
+Z        }
+        this.playbackBox.style.left = (gridSizeX * 8) * time;
+        this.playbackHandle.style.left = ((gridSizeX * 8) * time) - this.playbackHandleWidth;
+    }
+
+    remove() {
+        this.playbackBox.remove();
+        this.playbackHandle.remove();
+        this.playbackBox = null;
+        this.playbackHandle = null;
+    }
+}
+
 // object for handling the state/display of a measure, there will be four of these
 class Measure {
     constructor(score, number) {
@@ -385,8 +437,7 @@ class Measure {
         this.hoveredRow = -1;
         
         // playback state
-        this.playbackTime = -1;
-        this.playbackBox = null;
+        this.timingBar = null;
     }
 
     buildUI() {
@@ -411,6 +462,15 @@ class Measure {
             </div>
         `;
         this.container.appendChild(this.buttons);
+
+        // container for the timing bar
+        this.timingContainer = document.createElement("div");
+        this.timingContainer.className = "timingContainer";
+        this.timingContainer.style.width = (gridSizeX * 16) + "px";
+        this.timingContainer.style.height = (gridSizeY) + "px";
+        // back-reference
+        this.timingContainer.measure = this;
+        this.container.appendChild(this.timingContainer);
 
         // container for the background image and all note images
         this.imgContainer = document.createElement("div");
@@ -684,37 +744,23 @@ class Measure {
         this.score.doPlayback(button, [this], restart);
     }
 
-    startPlaybackMarker() {
+    startPlaybackMarker(playbackMarker) {
         // initialize the playback marker to just before the start
-        this.playbackTime = -0.1;
-
-        // create the marker if it's not present
-        if (this.playbackBox == null) {
-            // just a thin div
-            this.playbackBox = document.createElement("div");
-            // absolutely positioned
-            this.playbackBox.className = "playbackBox";
-            this.playbackBox.style.width = "4px";
-            this.playbackBox.style.height = (gridSizeY * 13) + "px";
-            this.playbackBox.style.left = 0;
-            this.playbackBox.style.top = 0;
-            // back-reference
-            this.playbackBox.measure = this;
-            this.imgContainer.appendChild(this.playbackBox);
-        }
+        this.playbackMarker = playbackMarker;
+        this.playbackMarker.setTime(this, -0.1);
     }
 
     setPlaybackMarkerTime(time) {
         // absolutely position the marker
-        this.playbackBox.style.left = (gridSizeX * 8) * time;
+        this.playbackMarker.setTime(this, -time);
 
         // re-initialize the playback time if it's wrapped around
-        if (time < this.playbackTime) {
-            this.playbackTime = -0.1;
+        if (time < this.playbackMarker.time) {
+            this.playbackMarker.time = -0.1;
         }
 
         // get the previous and current playback columns, 16 columns across a 2-second measure
-        var oldT = Math.floor(this.playbackTime * 8.0);
+        var oldT = Math.floor(this.playbackMarker.time * 8.0);
         var newT = Math.floor(time * 8.0);
 
         // loop from oldT + 1 to newT, if they are the same then this loops zero times,
@@ -730,7 +776,7 @@ class Measure {
         }
 
         // update state
-        this.playbackTime = time;
+        this.playbackMarker.setTime(this, time);
     }
 
     playAudioForTime(t, delay) {
@@ -760,9 +806,8 @@ class Measure {
 
     stopPlayback() {
         // clear playback state and UI
-        this.playbackTime = -1;
-        this.playbackBox.remove();
-        this.playbackBox = null;
+        this.playbackMarker.remove();
+        this.playbackMarker = null;
     }
 
     draw(context, imageMap, startX, startY, scale) {
@@ -1812,747 +1857,6 @@ class Score {
         var link = convertToPngLink(canvas, this.title);
         linkDiv.innerHTML = "";
         linkDiv.appendChild(link);
-    }
-}
-
-class Mixer {
-    constructor(score) {
-        // back reference to the score
-        this.score = score;
-
-        // build the UI
-        this.buildUI();
-    }
-
-    buildUI() {
-        // main container
-        this.mixerBox = document.createElement("div");
-        this.mixerBox.className = "mixerBox";
-        this.mixerBox.id = "mixerBox";
-        // back reference because why not
-        this.mixerBox.mixer = this;
-
-        this.container = this.mixerBox;
-
-        // button container
-        this.buttons = document.createElement("div");
-        this.buttons.className = "scoreButtonContainer";
-        // back-reference
-        this.buttons.score = this;
-        // build the buttons in a single row
-        this.buttons.innerHTML = `
-            <div class="scoreButtonRow">
-                <input class="titleButton" type="submit" value="Mixer"/>
-                <input class="button resetButton" type="submit" value="Reset"/>
-            </div>
-        `;
-        this.container.appendChild(this.buttons);
-
-        // hook up events
-        getFirstChild(this.container, "resetButton").addEventListener("click", () => { this.resetMixerButton() });
-        getFirstChild(this.container, "titleButton").addEventListener("click", () => { this.hide() });
-
-        // table fon containing the mixer list
-        var table = document.createElement("table");
-
-        // build a row slider
-        this.masterSlider = new MixerSlider(this, true, null, null, true);
-
-        // build the beginning of the mixer row
-        var tr = document.createElement("tr");
-        tr.className = "sectionRow";
-        tr.innerHTML = `<td style="text-align:right">Master</td><td/>`;
-
-        // add the slider and toggle
-        this.masterSlider.buildUI(tr);
-        // add the mixer row to the table
-        table.appendChild(tr);
-
-        // build the mixer UI for each section
-        for (var name in this.score.sections) {
-            this.score.sections[name].buildMixerUI(table);
-        }
-
-        // add the table to a scrollable container div
-        var tableDiv = document.createElement("div");
-        tableDiv.className = "mixerlistScollArea";
-        tableDiv.appendChild(table);
-
-        // add mixer list to container
-        this.container.appendChild(tableDiv);
-    }
-
-    init() {
-        // nothing to init
-    }
-
-    hide() {
-        toggleMixer(getFirstChild(this.container, "titleButton"));
-    }
-
-    resetMixerButton(e) {
-        // chrome thing
-        getFirstChild(this.container, "resetButton").blur();
-        this.resetMixer();
-    }
-
-    resetMixer() {
-        // reset each section
-        for (name in this.score.sections) {
-            this.score.sections[name].resetMixer();
-        }
-    }
-
-    mixerVolumeChange(isMixer, row, value, commit, secondary=false) {
-        // only called by the master slider
-        this.score.soundPlayer.setMasterVolume(value);
-    }
-
-
-    export() {
-        // moxer config code header
-        var string = "[Mixer";
-        var first = true;
-        var hasAdjustments = false;
-        // go over the sections
-        for (name in this.score.sections) {
-            // section delimiter
-            string += ":";
-            // get the section's config string
-            var sectionString = this.score.sections[name].exportMixer();
-            // add to the mixer config
-            string += sectionString;
-            // check if there actually was any section mixer config
-            if (sectionString.length > 0) {
-                hasAdjustments = true;
-            }
-        }
-
-        // if all sections are at default values then we don't need a mixer config at all
-        if (!hasAdjustments) {
-            return "";
-
-        } else {
-            // add the footer and return the config string
-            string += "]";
-            return string;
-        }
-    }
-
-    isMixerExportString(string) {
-        // check an import line for the mixer config format
-        return string.startsWith("[Mixer:") && string.endsWith("]");
-    }
-
-    import(string) {
-        // format check
-        if (!this.isMixerExportString(string)) {
-            throw "Invalid mixer format";
-        }
-        // extract the inside of the config and split by the section delimiter
-        var importSections = string.slice(7, -1).split(":");
-        // format check
-        // It's surprisingly hard to get the size of an associative array.
-        if (importSections.length != Object.getOwnPropertyNames(this.score.sections).length) {
-            throw "Invalid mixer format";
-        }
-
-        // iterate over the sections and config strings
-        var i = 0;
-        for (name in this.score.sections) {
-            // import mixer config into the section
-            this.score.sections[name].importMixer(importSections[i]);
-            i++;
-        }
-    }
-}
-
-function clearPlaylist(button) {
-    // chrome is doing strange things with clicked buttons so just unfocus it
-    button.blur();
-
-    // clear all menus
-    clearMenus();
-    // create a div and set some properties
-    var div = document.createElement("div");
-    div.className = "menu";
-    div.button = button;
-
-    // build the section menu out of buttons
-    var html = "";
-    html += `<input class="button" type="submit" value="Clear Playlist" onClick="reallyClearPlaylist(this)"/>`;
-    html += `<input class="button" type="submit" value="Cancel" onClick="clearMenus()"/>`;
-    div.innerHTML = html;
-
-    // put the menu in the clicked button's parent and anchor it to button
-    showMenu(div, getParent(button, "scoreButtonRow"), button);
-}
-
-function reallyClearPlaylist(button) {
-   // chrome is doing strange things with clicked buttons so just unfocus it
-    button.blur();
-
-    // need to get this before clearing menus
-    var playlist = getParent(button, "playlistBox").playlist;
-
-    // clear all menus
-    clearMenus();
-
-    // clear the playlist
-    playlist.clear();
-}
-
-function loopPlaylist(button) {
-    // chrome is doing strange things with clicked buttons so just unfocus it
-    button.blur();
-    var playlist = getParent(button, "playlistBox").playlist;
-
-    playlist.toggleLoop(button);
-}
-
-function addToPlaylist(button) {
-    // chrome is doing strange things with clicked buttons so just unfocus it
-    button.blur();
-    var playlist = getParent(button, "playlistBox").playlist;
-    playlist.add();
-}
-
-function editPlaylistSaveButton(e) {
-    editPlaylistSave(e.target);
-}
-
-function editPlaylistSave(button) {
-    var textarea = getFirstChild(getParent(button, "playlistBox"), "playlistEditArea");
-    if (textarea.exp != textarea.value) {
-        textarea.playlist.import(textarea.value);
-    }
-    clearMenus();
-    clearErrors();
-}
-
-function editPlaylist(button) {
-    // chrome is doing strange things with clicked buttons so just unfocus it
-    button.blur();
-    var playlist = getParent(button, "playlistBox").playlist;
-
-    // clear all menus
-    clearMenus();
-    // create a div and set some properties
-    var div = document.createElement("div");
-    div.className = "menu";
-    div.button = button;
-//    div.callback = callback
-
-    // build the section menu out of buttons
-    var textArea = document.createElement("textarea");
-    textArea.className = "playlistEditArea";
-    textArea.rows = "5";
-    textArea.cols = "64";
-    var exp = playlist.export();
-    textArea.value = exp;
-    textArea.exp = exp;
-    textArea.playlist = playlist;
-    // escape is usually ignored in text areas
-    textArea.onkeydown = textAreaKeyDown;
-
-    div.appendChild(textArea);
-
-    var save = document.createElement("input");
-    save.className = "button";
-    save.value = "Save";
-    save.textarea = textArea;
-    save.playlist = playlist;
-    save.onclick = editPlaylistSaveButton
-
-    div.appendChild(save);
-
-
-//    div.innerHTML = `<textarea id="songCode" rows="5" cols="64" onchange="editPlaylistChanged(this);"></textarea>`;
-
-    // put the menu in the clicked button's parent and anchor it to button
-    showMenu(div, getParent(button, "scoreButtonRow"), button);
-
-    textArea.focus();
-    textArea.select();
-}
-
-function textAreaKeyDown(e) {
-    e = e || window.event;
-    nodeName = e.target.nodeName;
-
-    switch (e.code) {
-		case "Escape" :
-		    // clear any open menus on escape
-		    clearMenu();
-		    break;
-		case "Enter" :
-		case "NumpadEnter" :
-		    // commit when enter is pressed
-		    editPlaylistSave(e.target);
-		    break;
-    }
-}
-
-class Playlist {
-    constructor(score) {
-        // back reference to the score
-        this.score = score;
-        // list of playlist entries
-        this.entries = Array();
-        // looping state
-        this.looping = false;
-        // build the UI
-        this.buildUI();
-    }
-
-    buildUI() {
-
-        // this.playlistContainer = document.createElement("div");
-        // this.playlistContainer.className = "playlistContainer";
-        // this.playlistContainer.measure = this;
-
-        // main container
-        // this will expand vertically with the playlist
-        // todo: figure out a good cross-browser way to add a scrollbar
-        this.playlistBox = document.createElement("div");
-        this.playlistBox.className = "playlistBox";
-        this.playlistBox.id = "playlistBox";
-        // back reference because why not
-        this.playlistBox.playlist = this;
-
-        // this.playlistContainer.appendChild(this.playlistBox);
-        this.playlistContainer = this.playlistBox;
-
-        // menu bar, just HTML it
-        this.playlistBox.innerHTML = `
-            <div class="scoreButtonRow">
-                <input class="titleButton" type="submit" value="Playlist"/>
-                <input class="button addButton" type="submit" value="Add" onClick="addToPlaylist(this)"/>
-                <input class="button loopButton" type="submit" value="Enable" onClick="loopPlaylist(this)"/>
-                <input class="button clearButton" type="submit" value="Clear" onClick="clearPlaylist(this)"/>
-                <input class="button editButton" type="submit" value="Copy/Paste" onClick="editPlaylist(this)"/>
-                <div class="popup">
-                    <input id="playlistCopyUrlButton" class="button urlButton popup" type="submit" value="Link"
-                           onClick="copyPlaylistUrl()"/>
-                    <div class="popuptext" id="playlistPopupBox">
-                        <input id="playlistUrlHolder" type="text" size="60" onblur="hideUrlPopup()"/>
-                    </div>
-                </div>
-            </div>
-        `;
-
-        // get a reference to the looping toggle button
-        this.loopingButton = getFirstChild(this.playlistBox, "loopButton");
-
-        getFirstChild(this.playlistBox, "titleButton").addEventListener("click", () => { this.hide() });
-
-        this.playlistScollArea = document.createElement("div");
-        this.playlistScollArea.className = "playlistScollArea";
-        this.playlistBox.appendChild(this.playlistScollArea);
-    }
-
-    hide() {
-        hidePlaylist();
-    }
-
-    addSongCode(code, select, insert=true) {
-        var song = new Song();
-        song.parseChatLink(code);
-        this.addSong(song, select, insert);
-    }
-
-    addSong(song, select, insert=true) {
-        // create a new entry
-        var entry = new PlaylistEntry(song, this);
-        if (this.selected && insert) {
-            // if there's a selection, the insert the new entry immediately after the selection
-            var index = this.entries.indexOf(this.selected);
-            this.entries.splice(index+1, 0, entry);
-            // do the same insertion in the dom
-            insertAfter(entry.playlistEntryContainer, this.selected.playlistEntryContainer);
-            // renumber entries
-            this.reIndex();
-
-        } else {
-            // no current selection or we're support to append instead of insert
-            this.entries.push(entry);
-            // no need to reindex the whole list
-            entry.setIndex(this.entries.length);
-            // insert into the dom
-            this.playlistScollArea.appendChild(entry.playlistEntryContainer);
-        }
-        if (select) {
-            // optionally select
-            this.select(entry, false);
-        }
-    }
-
-    removeEntry(entry) {
-        // remove from the entry list
-        if (removeFromList(this.entries, entry)) {
-            // remove from the dom
-            deleteNode(entry.playlistEntryContainer);
-            // if the node was selected then we have no selection now
-            if (this.selected == entry) {
-                this.selected = null;
-            }
-            // renumber entries
-            this.reIndex();
-        }
-    }
-
-    moveUp(entry) {
-        // get the index of the entry
-        var index = this.entries.indexOf(entry);
-        // make sure it's not already at the top
-        if (index > 0) {
-            // remove and insert one spot up
-            this.entries.splice(index, 1);
-            this.entries.splice(index-1, 0, entry);
-            // same move in the dom
-            deleteNode(entry.playlistEntryContainer);
-            insertBefore(entry.playlistEntryContainer, this.entries[index].playlistEntryContainer);
-            // renumber entries
-            this.reIndex();
-        }
-    }
-
-    moveToTop(entry) {
-        // get the index of the entry
-        var index = this.entries.indexOf(entry);
-        // make sure it's not already at the top
-        if (index > 0) {
-            // remove and insert at the beginning
-            this.entries.splice(index, 1);
-            this.entries.splice(0, 0, entry);
-            // same move in the dom
-            deleteNode(entry.playlistEntryContainer);
-            insertBefore(entry.playlistEntryContainer, this.entries[1].playlistEntryContainer);
-            // renumber entries
-            this.reIndex();
-        }
-    }
-
-    moveDown(entry) {
-        // get the index of the entry
-        var index = this.entries.indexOf(entry);
-        // make sure it's not already at the bottom
-        if (index >= 0 && index < this.entries.length - 1) {
-            // remove and insert one spot down
-            this.entries.splice(index, 1);
-            this.entries.splice(index+1, 0, entry);
-            // same move in the dom
-            deleteNode(entry.playlistEntryContainer);
-            insertAfter(entry.playlistEntryContainer, this.entries[index].playlistEntryContainer);
-            // renumber entries
-            this.reIndex();
-        }
-    }
-
-    moveToBottom(entry) {
-        // get the index of the entry
-        var index = this.entries.indexOf(entry);
-        if (index >= 0 && index < this.entries.length - 1) {
-            // remove and insert at the end
-            this.entries.splice(index, 1);
-            this.entries.splice(this.entries.length, 0, entry);
-            // same move in the dom
-            deleteNode(entry.playlistEntryContainer);
-            insertAfter(entry.playlistEntryContainer, this.entries[this.entries.length - 2].playlistEntryContainer);
-            // renumber entries
-            this.reIndex();
-        }
-    }
-
-    reIndex() {
-        // lazy, just loop through and re-index eveything.
-        for (var i = 0; i < this.entries.length; i++) {
-            this.entries[i].setIndex(i + 1);
-        }
-    }
-
-    select(entry, setScore, resetPlayback=false) {
-        // already selected
-        if (this.selected == entry) {
-            return;
-        }
-        // not sure if I need a null check but whatever
-        if (entry == null) {
-            return;
-        }
-        // check if there is already a selection
-        if (this.selected) {
-            // update the playlist entry's song, if this isn't an add action
-            if (setScore) {
-                this.selected.updateSong();
-            }
-            // clear selection
-            this.selected.setSelected(false);
-        }
-        // select the new entry
-        entry.setSelected(true);
-        this.selected = entry;
-        // update the score, if this isn't an add action
-        if (setScore) {
-            this.score.setSongObject(this.selected.song, true, resetPlayback);
-            // let's make this simple for now: switching songs in the playlist clears the undo history.
-            clearUndoStack();
-        }
-    }
-
-    selectNext() {
-        // if there is no selection then select the first entry
-        if (!this.selected) {
-            this.select(this.entries[0], true);
-            return;
-        }
-        // get the currently selected index
-        var index = this.entries.indexOf(this.selected);
-        // increment and wrap around if necessary
-        index += 1;
-        if (index >= this.entries.length) {
-            index = 0;
-        }
-        // change the selection, updating the score
-        this.select(this.entries[index], true);
-    }
-
-    clear() {
-        // delete from the dom
-        for (var i = 0; i < this.entries.length; i++) {
-            deleteNode(this.entries[i].playlistEntryContainer);
-        }
-        // clear state
-        this.entries = Array();
-        this.selected = null;
-    }
-
-
-    incrementName(name) {
-        var incrementNameRegex = /(.*?) (\d+)/;
-        var match = name.match(incrementNameRegex);
-        if (match) {
-            return match[1] + " " + (parseInt(match[2]) + 1);
-        } else {
-            return name + " " + 2;
-        }
-    }
-
-    add() {
-        // make the name unique
-        var name = this.score.title;
-        for (;;) {
-            if (!this.entries.find((entry) => entry.song.name == name)) {
-                break;
-            }
-            name = this.incrementName(name);
-        }
-        if (name != this.score.title) {
-            this.score.setTitle(name, false);
-        }
-        // add the song currently in the score and automatically select it
-        // This bypasses the auto-update when the selection changes, so the previously selected entry remains unchanged
-        this.addSong(this.score.getSongObject(), true);
-    }
-
-    toggleLoop() {
-        this.setLooping(!this.looping)
-    }
-
-    setLooping(looping) {
-        if (!looping) {
-            this.loopingButton.value = "Enable";
-            this.looping = false;
-        } else {
-            this.loopingButton.value = "Disable";
-            this.looping = true;
-        }
-    }
-
-    export() {
-        // build a string with each playlist song's code on a new line
-        var str = "";
-        for (var i = 0; i < this.entries.length; i++) {
-            str = str + this.entries[i].song.formatAsChatLink() + "\n";
-        }
-        // see if there's a mixer config to export
-        var mixerString = this.score.mixer.export();
-        if (mixerString != "") {
-            // put the mixer config at the end
-            str = str + mixerString + "\n";
-        }
-        return str;
-    }
-
-    import(str) {
-        // split into lines
-        var songCodes = str.split("\n")
-        var readEntries = Array();
-
-        var hasMixerConfig = false;
-        for (var i = 0; i < songCodes.length; i++) {
-            // trim and check for blank lines
-            var code = songCodes[i].trim();
-
-            // check for a mixer config
-            if (this.score.mixer.isMixerExportString(code)) {
-                // make sure the mixer is visible
-                showMixer();
-                // import the mixer config
-                this.score.mixer.import(code);
-                hasMixerConfig = true;
-
-            } else if (code != "") {
-                // parse the song
-                var song = new Song();
-                song.parseChatLink(code);
-                readEntries.push(song);
-            }
-
-            if (!hasMixerConfig) {
-                this.score.mixer.resetMixer();
-                hideMixer();
-            }
-        }
-
-        // don't make any changes if we didn't read any valid songs
-        // we also avoid making any changes if there was an error parsing the song list
-        if (readEntries.length > 0) {
-            // clear the current playlist
-            this.clear();
-            // add each song
-            for (var i = 0; i < readEntries.length; i++) {
-                // add it to the playlist, without affecting the selection
-                this.addSong(readEntries[i], false);
-            }
-            // finally, select the first entry and reset playback
-            this.select(this.entries[0], true, true);
-        }
-    }
-}
-
-class PlaylistEntry {
-    constructor(song, playlist) {
-        // song object
-        this.song = song;
-        // back reference
-        this.playlist = playlist;
-        // build the UI
-        this.buildUI();
-    }
-
-    buildUI() {
-        // main container
-        this.playlistEntryContainer = document.createElement("div");
-        this.playlistEntryContainer.className = "playlistEntryContainer";
-        this.playlistEntryContainer.playlist = this;
-
-        // I kind of regret this, but build the dom manually
-        {
-            var span = document.createElement("span");
-            span.className = "smallButton";
-            span.onclick = this.deletePlaylistEntry
-            span.entry = this;
-            span.innerHTML = `X`;
-            this.playlistEntryContainer.appendChild(span);
-        }
-
-        {
-            var span = document.createElement("span");
-            span.className = "smallButton";
-            span.onclick = this.movePlaylistEntryToTop
-            span.entry = this;
-            span.innerHTML = `⇈`;
-            this.playlistEntryContainer.appendChild(span);
-        }
-        {
-            var span = document.createElement("span");
-            span.className = "smallButton";
-            span.onclick = this.movePlaylistEntryUp
-            span.entry = this;
-            span.innerHTML = `↑`;
-            this.playlistEntryContainer.appendChild(span);
-        }
-
-        {
-            var span = document.createElement("span");
-            span.className = "smallButton";
-            span.onclick = this.movePlaylistEntryDown
-            span.entry = this;
-            span.innerHTML = `↓`;
-            this.playlistEntryContainer.appendChild(span);
-        }
-        {
-            var span = document.createElement("span");
-            span.className = "smallButton";
-            span.onclick = this.movePlaylistEntryToBottom
-            span.entry = this;
-            span.innerHTML = `⇊`;
-            this.playlistEntryContainer.appendChild(span);
-        }
-
-        {
-            // we need to keep a reference to the index span to change its color when selected
-            this.indexBar = document.createElement("span");
-            this.indexBar.className = "playlistEntry";
-            this.indexBar.onclick = this.selectPlaylistEntry
-            this.indexBar.entry = this;
-            this.playlistEntryContainer.appendChild(this.indexBar);
-        }
-        {
-            // we need to keep a reference to the title span to change its color when selected
-            this.titleBar = document.createElement("span");
-            this.titleBar.className = "playlistEntry";
-            this.titleBar.onclick = this.selectPlaylistEntry
-            this.titleBar.entry = this;
-            this.titleBar.innerHTML = this.song.getName();
-            this.playlistEntryContainer.appendChild(this.titleBar);
-        }
-    }
-
-    setIndex(index) {
-        this.indexBar.innerHTML = index;
-    }
-
-    deletePlaylistEntry() {
-        this.entry.playlist.removeEntry(this.entry);
-    }
-
-    movePlaylistEntryUp() {
-        this.entry.playlist.moveUp(this.entry);
-    }
-
-    movePlaylistEntryToTop() {
-        this.entry.playlist.moveToTop(this.entry);
-    }
-
-    movePlaylistEntryDown() {
-        this.entry.playlist.moveDown(this.entry);
-    }
-
-    movePlaylistEntryToBottom() {
-        this.entry.playlist.moveToBottom(this.entry);
-    }
-
-    selectPlaylistEntry() {
-        this.entry.playlist.select(this.entry, true);
-    }
-
-    setSelected(selected) {
-        // change the css depending on whether it's selected
-        this.indexBar.className = selected ? "playlistEntrySelected" : "playlistEntry";
-        this.titleBar.className = selected ? "playlistEntrySelected" : "playlistEntry";
-        if (selected) {
-            // scroll the playlist viewer to the selected entry, either at the top or the botton, whichever is nearest
-            this.playlistEntryContainer.scrollIntoView({"behavior": "auto", "block": "nearest", "inline": "nearest"});
-        }
-    }
-
-    updateSong() {
-        // load the current song from the score
-        this.song = this.playlist.score.getSongObject();
-        // update the title
-        this.titleBar.innerHTML = this.song.getName();
     }
 }
